"Add date-and-time package and update HomeAttendancePage to use it"
This commit adds the "date-and-time" package to the project and updates the HomeAttendancePage component to use it to format the current date and time. The package is added to both package.json and package-lock.json, and the HomeAttendancePage component is updated to import and use the package to format the current date and time. The commit also updates the test scripts to account for the changes made to the HomeAttendancePage component.